% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.R
\name{mw_standardize}
\alias{mw_standardize}
\title{Standardize healthcare wait time data}
\usage{
mw_standardize(df)
}
\arguments{
\item{df}{Raw data frame retrieved from the API}
}
\value{
A standardized tibble
}
\description{
Standardize healthcare wait time data
}
